fixed size warning

+++ src/cef/RequestClient.h
@@ -13,7 +13,8 @@ public:
           UNREFERENCED_PARAMETER(request);
 
           CefResponse::HeaderMap headers;
-          response->status = new HttpStatus(request->GetResponse()->GetStatus(), request->GetResponse()->GetStatusText().ToString());
+          response->status = new HttpStatus((uint16_t)request->GetResponse()->GetStatus(), 
+                                            request->GetResponse()->GetStatusText().ToString());
           request->GetResponse()->GetHeaderMap(headers);
           for (auto it = headers.begin(); it != headers.end(); it++) {
                response->Headers().Add(std::wstring(it->first), std::wstring(it->second));
